S20 delete image when update new image or delete product

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 // const Mongoose = require('mongoose')
+const fileHelper = require('../util/file');
 
 const { validationResult } = require('express-validator/check');
 const Product = require('../models/product');
@@ -157,6 +158,7 @@ exports.postEditProduct = (req, res, next) => {
       }
       Object.assign(product, { title, description, price });
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save()
@@ -196,7 +198,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error("Product not found"))
+      }
+      fileHelper.deleteFile(product.imageUrl)
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(() => {
       console.log('destroy product')
       res.redirect('/admin/products');
@@ -206,4 +215,5 @@ exports.postDeleteProduct = (req, res, next) => {
       error.httpStatisCode = 500
       return next(error)
     });
+
 }
